fix an issue where window sizes are restored improperly
This commit construct commands to restore window sizes without winrestcmd().

The result of winrestcmd() includes commands for floating windows in Neovim.
Windows sizes are saved before windows of quickpick open, and
the floating windows close after the cursor moves to a window of quickpick.
The number of windows, therefore, changed between when winrestcmd() was called
and when the result of winrestcmd() was executed.